Extract avro macros in their own submodule
