+++ scio-avro-base/src/main/scala/com/spotify/scio/avro/syntax/Syntax.scala
