+++ scio-avro-typed/src/it/scala/com/spotify/scio/avro/types/AvroTypeIT.scala
