+++ scio-avro-typed/src/main/scala/com/spotify/scio/avro/AvroTyped.scala
@@ -1,0 +1,94 @@
+/*
+ * Copyright 2019 Spotify AB.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package com.spotify.scio.avro
+
+import com.spotify.scio.avro.types.AvroType.HasAvroAnnotation
+import com.spotify.scio.coders.Coder
+import com.spotify.scio.io._
+import com.spotify.scio.util.ScioUtil
+import com.spotify.scio.values._
+import com.spotify.scio.{avro, ScioContext}
+import org.apache.avro.file.CodecFactory
+import org.apache.avro.generic.GenericRecord
+import org.apache.beam.sdk.transforms.SerializableFunction
+import org.apache.beam.sdk.{io => beam}
+
+import scala.jdk.CollectionConverters._
+import scala.reflect.runtime.universe._
+
+object AvroTyped {
+  final case class AvroIO[T <: HasAvroAnnotation: TypeTag: Coder](path: String) extends ScioIO[T] {
+    override type ReadP = Unit
+    override type WriteP = avro.AvroIO.WriteParam
+    final override val tapT: TapT.Aux[T, T] = TapOf[T]
+
+    private def typedAvroOut[U](
+      write: beam.AvroIO.TypedWrite[U, Void, GenericRecord],
+      path: String,
+      numShards: Int,
+      suffix: String,
+      codec: CodecFactory,
+      metadata: Map[String, AnyRef]
+    ) =
+      write
+        .to(ScioUtil.pathWithShards(path))
+        .withNumShards(numShards)
+        .withSuffix(suffix)
+        .withCodec(codec)
+        .withMetadata(metadata.asJava)
+
+    /**
+     * Get a typed SCollection from an Avro schema.
+     *
+     * Note that `T` must be annotated with
+     * [[com.spotify.scio.avro.types.AvroType AvroType.fromSchema]],
+     * [[com.spotify.scio.avro.types.AvroType AvroType.fromPath]], or
+     * [[com.spotify.scio.avro.types.AvroType AvroType.toSchema]].
+     */
+    override protected def read(sc: ScioContext, params: ReadP): SCollection[T] = {
+      val avroT = AvroType[T]
+      val t = beam.AvroIO.readGenericRecords(avroT.schema).from(path)
+      sc.applyTransform(t).map(avroT.fromGenericRecord)
+    }
+
+    /**
+     * Save this SCollection as an Avro file. Note that element type `T` must be a case class
+     * annotated with [[com.spotify.scio.avro.types.AvroType AvroType.toSchema]].
+     */
+    override protected def write(data: SCollection[T], params: WriteP): Tap[T] = {
+      val avroT = AvroType[T]
+      val t = beam.AvroIO
+        .writeCustomTypeToGenericRecords()
+        .withFormatFunction(new SerializableFunction[T, GenericRecord] {
+          override def apply(input: T): GenericRecord =
+            avroT.toGenericRecord(input)
+        })
+        .withSchema(avroT.schema)
+      data.applyInternal(
+        typedAvroOut(t, path, params.numShards, params.suffix, params.codec, params.metadata)
+      )
+      tap(())
+    }
+
+    override def tap(read: ReadP): Tap[T] = {
+      val avroT = AvroType[T]
+      GenericRecordTap(ScioUtil.addPartSuffix(path), avroT.schema)
+        .map(avroT.fromGenericRecord)
+    }
+  }
+}
