+++ scio-avro-typed/src/main/scala/com/spotify/scio/avro/syntax/ScioContextTypedSyntax.scala
@@ -1,0 +1,49 @@
+/*
+ * Copyright 2019 Spotify AB.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package com.spotify.scio.avro.syntax
+
+import com.spotify.scio.ScioContext
+import com.spotify.scio.avro._
+import com.spotify.scio.avro.types.AvroType.HasAvroAnnotation
+import com.spotify.scio.coders.Coder
+import com.spotify.scio.values._
+
+import scala.reflect.runtime.universe._
+
+/** Enhanced version of [[ScioContext]] with Avro methods. */
+final class ScioContextTypedOps(private val self: ScioContext) extends AnyVal {
+
+  /**
+   * Get a typed SCollection from an Avro schema.
+   *
+   * Note that `T` must be annotated with
+   * [[com.spotify.scio.avro.types.AvroType AvroType.fromSchema]],
+   * [[com.spotify.scio.avro.types.AvroType AvroType.fromPath]], or
+   * [[com.spotify.scio.avro.types.AvroType AvroType.toSchema]].
+   */
+  def typedAvroFile[T <: HasAvroAnnotation: TypeTag: Coder](
+    path: String
+  ): SCollection[T] =
+    self.read(AvroTyped.AvroIO[T](path))
+}
+
+/** Enhanced with Avro methods. */
+trait ScioContextTypedSyntax {
+  implicit def avroScioContextTypedOps(c: ScioContext): ScioContextTypedOps =
+    new ScioContextTypedOps(c)
+}
