+++ scio-avro-typed/src/main/scala/com/spotify/scio/avro/types/SchemaUtil.scala
