+++ scio-avro-typed/src/test/avro/scio-avro-test.avsc
