+++ scio-avro-typed/src/test/scala/com/spotify/scio/avro/types/SchemaUtilTest.scala
